[tests] Shared func
