+++ communicator/communicator_test.go
@@ -1,7 +1,6 @@
 package communicator
 
 import (
-	"github.com/spf13/viper"
 	config "github.com/weldpua2008/suprasched/config"
 
 	"errors"
@@ -40,35 +39,35 @@ func TestGetCommunicator(t *testing.T) {
 	}
 }
 
-func loadCfg(t *testing.T, CfgFile string) {
-	tmp := config.C
-	defer func() {
-		config.C = tmp
-	}()
-	config.C = config.Config{}
-	config.CfgFile = CfgFile
-	viper.SetConfigFile(config.CfgFile)
-	// t.Logf("Loaded: %v", config.CfgFile)
-	if err := viper.ReadInConfig(); err != nil {
-		t.Errorf("Can't read config: %v", err)
-	}
-	err := viper.Unmarshal(&config.C)
-	if err != nil {
-		t.Errorf("unable to decode into struct, %v", err)
-
-	}
-
-	if config.C.ClientId == string("") {
-		t.Errorf("Expected C.ClientId not empty got %v\n", config.C)
-	}
-	if config.C.ConfigVersion == string("") {
-		t.Errorf("Expected C.ConfigVersion not empty got %v\n", config.C)
-	}
-
-}
+// func loadCfg(t *testing.T, CfgFile string) {
+// 	tmp := config.C
+// 	defer func() {
+// 		config.C = tmp
+// 	}()
+// 	config.C = config.Config{}
+// 	config.CfgFile = CfgFile
+// 	viper.SetConfigFile(config.CfgFile)
+// 	// t.Logf("Loaded: %v", config.CfgFile)
+// 	if err := viper.ReadInConfig(); err != nil {
+// 		t.Errorf("Can't read config: %v", err)
+// 	}
+// 	err := viper.Unmarshal(&config.C)
+// 	if err != nil {
+// 		t.Errorf("unable to decode into struct, %v", err)
+//
+// 	}
+//
+// 	if config.C.ClientId == string("") {
+// 		t.Errorf("Expected C.ClientId not empty got %v\n", config.C)
+// 	}
+// 	if config.C.ConfigVersion == string("") {
+// 		t.Errorf("Expected C.ConfigVersion not empty got %v\n", config.C)
+// 	}
+//
+// }
 
 func TestGetSectionCommunicator(t *testing.T) {
-	loadCfg(t, "fixtures/http.yml")
+	config.LoadCfgForTests(t, "fixtures/http.yml")
 
 	cases := []struct {
 		section string
